+++ examples/src/test/scala/commbank/coppersmith/examples/userguide/JoinFeaturesSpec.scala
@@ -53,7 +53,7 @@ object JoinFeaturesSpec extends ThermometerHiveSpec { def is = s2"""
 
     executesSuccessfully(job.job) must_== JobFinished
 
-    val outPath = s"$dir/user/dev/view/warehouse/features/ratings/year=2015/month=01/day=01/*"
+    val outPath = s"$dir/user/dev/ratings/year=2015/month=01/day=01/*"
     expectations { context =>
       context.lines(new Path(outPath)).toSet must_==
         Set(s"2|$featureName|3.5|2015-01-01")