+++ scalding/src/main/scala/commbank/coppersmith/scalding/EavtText.scala
@@ -14,8 +14,6 @@
 
 package commbank.coppersmith.scalding
 
-import commbank.coppersmith.scalding.HiveSupport.{FailJob, DelimiterConflictStrategy}
-import org.apache.hadoop.fs.Path
 import org.joda.time.DateTime
 
 import au.com.cba.omnia.maestro.api._, Maestro._
@@ -23,14 +21,10 @@ import au.com.cba.omnia.maestro.api._, Maestro._
 import commbank.coppersmith.{FeatureValue, Feature}, Feature._, Value._
 import commbank.coppersmith.thrift.Eavt
 
-object EavtTextSink {
-  import TextSink._
-
-  val defaultPartition = DerivedSinkPartition[Eavt, (String, String, String)](
+object EavtText {
+  val eavtByDay = DerivedSinkPartition[Eavt, (String, String, String)](
     HivePartition.byDay(Fields[Eavt].Time, "yyyy-MM-dd")
   )
-
-  type EavtTextSink = TextSink[Eavt]
 
   implicit object EavtEnc extends FeatureValueEnc[Eavt] {
     def encode(fvt: (FeatureValue[_], Time)): Eavt = fvt match {
@@ -46,14 +40,4 @@ object EavtTextSink {
         Eavt(fv.entity, fv.name, featureValue, featureTime)
     }
   }
-
-  def configure(dbPrefix:  String,
-                 dbRoot:    Path,
-                 tableName: TableName,
-                 partition: SinkPartition[Eavt] = defaultPartition,
-                 group:     Option[String] = None,
-                 dcs:       DelimiterConflictStrategy[Eavt] = FailJob[Eavt]()): EavtTextSink = {
-    TextSink.configure(dbPrefix, dbRoot, tableName, partition, group, dcs)
-  }
-
 }